Peers download genesis block from trusted peer, move peer utitilities functions
