+++ crates/chain/src/lib.rs
@@ -1,5 +1,6 @@
 pub(crate) mod arbiter_handle;
 pub mod chain;
 pub mod vdf;
+pub mod peer_utilities;
 
 pub use chain::*;
